chore(tests): move ParseRemainingIntoTableDirect to plugin_collection
