+++ falco_plugin_tests/src/plugin_collection/mod.rs
@@ -1,2 +1,3 @@
 pub mod extract;
+pub mod parse;
 pub mod source;
