+++ falco_plugin_tests/src/plugin_collection/parse/mod.rs
@@ -1,0 +1,1 @@
+pub mod remaining_into_table_direct;
